# secureprompt-schemas/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: SecurePrompt Gateway API
  version: "1.0.0"
  description: >
    SecurePrompt is an LLM security gateway that sits between enterprise
    applications and AI providers. It redacts PII/secrets, enforces policy,
    tracks token costs, and provides governance analytics.

    Two auth schemes are used:
    - **JWT** (`Authorization: Bearer eyJ...`) — dashboard routes
    - **API Key** (`Authorization: Bearer sp-...`) — gateway + MCP utility routes
servers:
  - url: http://localhost:8080
    description: Local dev Rust gateway
  - url: https://api.secureprompt.tech
    description: Cloud
...
paths:
...
  /v1/auth/token:
    post:
      tags: [Auth]
      summary: Issue access + refresh tokens for email/password
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New tokens issued (refresh-token rotation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Revoked

  # ── Gateway (OpenAI-compatible) ───────────────────────────────────────────────
  /v1/chat/completions:
    post:
      tags: [Gateway]
      summary: Chat completions (OpenAI-compatible)
      description: >
        Proxies to the configured LLM provider after PII redaction and policy
        enforcement. Supports streaming via `stream: true`. Use an API key
        (`sp-...`) in the Authorization header.
      operationId: chatCompletions
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionsRequest"
      responses:
        "200":
          description: >
            OpenAI-compatible chat completion response, or a server-sent event
            stream when `stream: true`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionsResponse"
            text/event-stream:
              schema:
                type: string
                description: "SSE stream of data: {...} chunks"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/BudgetExceeded"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/completions:
    post:
      tags: [Gateway]
      summary: Text completions — legacy OpenAI endpoint
      operationId: completions
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionRequest"
      responses:
        "200":
          description: OpenAI-compatible completion response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/BudgetExceeded"
  /v1/embeddings:
    post:
      tags: [Gateway]
      summary: Embeddings (OpenAI-compatible)
      operationId: embeddings
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbeddingsRequest"
      responses:
        "200":
          description: OpenAI-compatible embeddings response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/BudgetExceeded"

  # ── MCP Utilities ─────────────────────────────────────────────────────────────
  /v1/redact:
    post:
      tags: [MCP]
      summary: Redact PII and secrets from text
      description: >
        Runs regex + ML NER detection on the input text, replaces detected
        values with opaque tokens, and returns the redacted text along with
        the detection list. Respects workspace policy rules (may deny or
        further redact). Requires an API key.
      operationId: redact
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactRequest"
      responses:
        "200":
          description: Redacted text with detection list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedactResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/tokens/estimate:
    post:
      tags: [MCP]
      summary: Estimate token count for a text + model pair
      description: >
        Approximates token count (1 token ≈ 4 characters, GPT-family
        heuristic). No auth required for this endpoint.
      operationId: estimateTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRequest"
      responses:
        "200":
          description: Token estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResponse"
  /v1/policy/check:
    post:
      tags: [MCP]
      summary: Dry-run policy evaluation (no provider call)
      description: >
        Evaluates workspace policy rules against the given text without
        making a downstream provider call. Returns whether the request would
        be allowed and which policy events fired. Requires an API key.
      operationId: checkPolicy
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCheckRequest"
      responses:
        "200":
          description: Policy check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyCheckResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Analytics ─────────────────────────────────────────────────────────────────
...
  /v1/requests:
    get:
      tags: [Requests]
      summary: List gateway requests (cursor-paginated)
      description: >
        Reads raw `request_events` + `policy_events` from ClickHouse.
        Maximum window: 90 days. Maximum page size: 200.
        Pagination: pass `next_cursor` from the previous response as `cursor`.
      operationId: listRequests
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
          description: Start of time window (default now - 7 days; max 90 days ago)
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
          description: End of time window (default now)
        - in: query
          name: model
          required: false
          schema:
            type: string
        - in: query
          name: has_violation
          required: false
          schema:
            type: boolean
          description: Filter to only requests that triggered at least one policy event
        - in: query
          name: cursor
          required: false
          schema:
            type: string
          description: Opaque cursor from previous response's `next_cursor`
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: "#/components/parameters/WorkspaceIdQuery"
      responses:
        "200":
          description: Paginated request list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRequestsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/keys:
    get:
      tags: [Keys]
      summary: List API keys for the workspace
      operationId: listKeys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of API keys (plaintext never included)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Keys]
      summary: Create a new API key (admin only)
      description: >
        Returns the full plaintext key exactly once in `api_key`.
        Store it immediately — it cannot be retrieved again.
      operationId: createKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeyRequest"
      responses:
        "201":
          description: Key created — plaintext returned once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/providers:
    get:
      tags: [Providers]
      summary: List LLM providers
      operationId: listProviders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of providers (credential never included)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Providers]
      summary: Create a provider (admin only)
      operationId: createProvider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/policy-rules:
    get:
      tags: [PolicyRules]
      summary: List policy rules
      operationId: listPolicyRules
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of policy rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyRuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [PolicyRules]
      summary: Create a policy rule (admin only)
      description: >
        Action must be one of: `deny`, `allow`, `redact`, `transform`, `flag`.
        Priority must be unique within the workspace (409 on conflict).
        The rule is immediately indexed in Qdrant for semantic policy lookup.
      operationId: createPolicyRule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Priority already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriorityConflict"
...
  /v1/telemetry/client-error:
    post:
      tags: [Telemetry]
      summary: Report a frontend error event
      description: >
        Accepts bounded frontend error events. Logs via structured tracing
        and increments the Prometheus counter
        `secureprompt_dashboard_client_errors_total{component}`.
        Does NOT persist to Postgres or ClickHouse.
      operationId: reportClientError
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientErrorEvent"
      responses:
        "204":
          description: Event accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Billing ───────────────────────────────────────────────────────────────────
...
  /v1/secure-mode:
    get:
      tags: [Settings]
      summary: Get secure mode configuration
      operationId: getSecureMode
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current secure mode configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureMode"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Settings]
      summary: Update secure mode configuration
      operationId: putSecureMode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecureModeRequest"
      responses:
        "200":
          description: Secure mode configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureMode"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Dashboard JWT — Authorization: Bearer eyJ..."
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API key — Authorization: Bearer sp-..."
  parameters:
...
    WorkspaceIdQuery:
      in: query
      name: workspace_id
      required: false
      schema:
        type: string
        format: uuid
      description: IDOR guard — must equal the JWT workspace_id when provided
  responses:
    BadRequest:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Insufficient role or workspace mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    BudgetExceeded:
      description: Workspace budget exhausted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BudgetExceededError"
  schemas:
...
    TokenRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 1
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - access_expires_at
        - refresh_expires_at
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        access_expires_at:
          type: integer
          format: int64
          description: Unix seconds
        refresh_expires_at:
          type: integer
          format: int64
          description: Unix seconds
        user:
          type: object
          required: [id, email]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
        workspace_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [admin, member, viewer]

    # ── Gateway ─────────────────────────────────────────────────────────────────
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
    ChatCompletionsRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          example: gpt-4o
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
          default: false
        stream_options:
          type: object
          nullable: true
          properties:
            include_usage:
              type: boolean
    ChatCompletionsResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                $ref: "#/components/schemas/ChatMessage"
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
    CompletionRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          description: String or array of strings
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        stream:
          type: boolean
          default: false
    EmbeddingsRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
          example: text-embedding-3-small
        input:
          description: String or array of strings
          oneOf:
            - type: string
            - type: array
              items:
                type: string

    # ── MCP ─────────────────────────────────────────────────────────────────────
    RedactRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
    DetectionItem:
      type: object
      required: [class, value]
      properties:
        class:
          type: string
          example: EMAIL
        value:
          type: string
          example: john@example.com
    RedactResponse:
      type: object
      required: [redacted_text, detections, vault_size]
      properties:
        redacted_text:
          type: string
        detections:
          type: array
          items:
            $ref: "#/components/schemas/DetectionItem"
        vault_size:
          type: integer
          minimum: 0
    EstimateRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        model:
          type: string
          nullable: true
    EstimateResponse:
      type: object
      required: [estimated_tokens, model]
      properties:
        estimated_tokens:
          type: integer
          format: int64
          minimum: 0
        model:
          type: string
    PolicyCheckRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        model:
          type: string
          nullable: true
    PolicyCheckResponse:
      type: object
      required: [allowed, action, events]
      properties:
        allowed:
          type: boolean
        action:
          type: string
          example: deny
        events:
          type: array
          items:
            type: string

    # ── Requests ────────────────────────────────────────────────────────────────
    RequestListItem:
      type: object
      required:
        - request_id
        - workspace_id
        - provider
        - model
        - final_action
        - cost_usd
        - has_violation
        - created_at
      properties:
        request_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        provider:
          type: string
        model:
          type: string
        final_action:
          type: string
        input_tokens:
          type: integer
          nullable: true
        output_tokens:
          type: integer
          nullable: true
        cost_usd:
          type: number
          format: double
        has_violation:
          type: boolean
        created_at:
          type: string
          format: date-time
    ListRequestsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RequestListItem"
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page; null when no more pages
...
    CreateKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    KeyResponse:
      type: object
      required: [id, name, prefix, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
          description: First 8 characters of the key (safe to display)
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
    CreateKeyResponse:
      type: object
      required: [id, name, api_key, prefix, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        api_key:
          type: string
          description: Full plaintext key — shown once, never stored
        prefix:
          type: string
        created_at:
          type: string
          format: date-time
...
    ProviderResponse:
      type: object
      required: [id, name, provider_type, has_credential, last_rotated_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        provider_type:
          type: string
          example: openai
        has_credential:
          type: boolean
          description: True when an encrypted credential is stored
        last_rotated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateProviderRequest:
      type: object
      required: [name, provider_type]
      properties:
        name:
          type: string
        provider_type:
          type: string
          example: openai
        credential:
          type: string
          nullable: true
          description: Plaintext API key — encrypted before storage
...
    PolicyAction:
      type: string
      enum: [deny, allow, redact, transform, flag]
    PolicyRuleResponse:
      type: object
      required:
        - id
        - workspace_id
        - name
        - priority
        - conditions
        - action
        - action_params
        - enabled
        - dry_run
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        priority:
          type: integer
          description: Lower = higher precedence; must be unique per workspace
        conditions:
          type: object
          description: Rule conditions (arbitrary JSON)
        action:
          $ref: "#/components/schemas/PolicyAction"
        action_params:
          type: object
          description: Action parameters (arbitrary JSON)
        enabled:
          type: boolean
        dry_run:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRuleRequest:
      type: object
      required: [name, priority, action]
      properties:
        name:
          type: string
        priority:
          type: integer
        conditions:
          type: object
          nullable: true
        action:
          $ref: "#/components/schemas/PolicyAction"
        action_params:
          type: object
          nullable: true
        enabled:
          type: boolean
          default: true
        dry_run:
          type: boolean
          default: false
...
    PriorityConflict:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: priority_conflict
        message:
          type: string

    # ── Telemetry ───────────────────────────────────────────────────────────────
    ClientErrorEvent:
      type: object
      required: [component, message, url]
      properties:
        component:
          type: string
          maxLength: 128
        message:
          type: string
          maxLength: 1024
        url:
          type: string
          maxLength: 512
        user_agent:
          type: string
          nullable: true
        context:
          type: object
          nullable: true
          description: Arbitrary structured context; max 2 KB serialized

    # ── Workspaces / Budgets ────────────────────────────────────────────────────
...
    SecureMode:
      type: object
      required:
        - enabled
        - level
        - block_on_pii_detection
        - block_on_injection_detection
        - redact_pii_in_responses
        - updated_at
      properties:
        enabled:
          type: boolean
        level:
          type: string
          enum: [permissive, standard, strict]
        block_on_pii_detection:
          type: boolean
        block_on_injection_detection:
          type: boolean
        redact_pii_in_responses:
          type: boolean
        updated_at:
          type: string
          format: date-time
    SecureModeRequest:
      type: object
      properties:
        enabled:
          type: boolean
        level:
          type: string
          enum: [permissive, standard, strict]
        block_on_pii_detection:
          type: boolean
        block_on_injection_detection:
          type: boolean
        redact_pii_in_responses:
          type: boolean
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
              example: secureprompt_error
            code:
              type: string
              description: Stable machine-readable code (e.g. budget_exceeded)
    BudgetExceededError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, type]
          properties:
            code:
              type: string
              example: budget_exceeded
            message:
              type: string
            type:
              type: string
              example: secureprompt_error